Refactor Extensions
