+++ src/main.js
@@ -1,7 +1,8 @@
 import Vue from 'vue';
 import ModelerApp from './ModelerApp.vue';
 import BootstrapVue from 'bootstrap-vue';
-import './eventBus';
+import './setup/initialLoad';
+
 
 Vue.use(BootstrapVue);
