+++ src/setup/eventBus.js
@@ -1,0 +1,3 @@
+import Vue from 'vue';
+
+window.ProcessMaker = { EventBus: new Vue() };
