+++ src/setup/initialLoad.js
@@ -1,0 +1,63 @@
+import './eventBus';
+
+/* Expression example */
+import bpmnExtension from '@processmaker/processmaker-bpmn-moddle/resources/processmaker.json';
+
+// Import Extensions for Testing Modeler
+import './extensions/loadUsers';
+import './extensions/testTaskInspectorExtension';
+import './extensions/twitterConnector';
+
+/* Our initial node types to register with our modeler */
+import {
+  association,
+  endEvent,
+  exclusiveGateway,
+  inclusiveGateway,
+  parallelGateway,
+  sequenceFlow,
+  startEvent,
+  task,
+  textAnnotation,
+  pool,
+  poolLane,
+} from '../components/nodes';
+
+const nodeTypes = [
+  startEvent,
+  endEvent,
+  task,
+  exclusiveGateway,
+  inclusiveGateway,
+  parallelGateway,
+  sequenceFlow,
+  textAnnotation,
+  association,
+  pool,
+  poolLane,
+];
+
+const blank = `
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" id="Definitions_03dabax" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="2.0.3">
+  <bpmn:process id="Process_1" isExecutable="true"></bpmn:process>
+  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Process_1"></bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+</bpmn:definitions>
+`;
+
+window.ProcessMaker.EventBus.$on('modeler-init', modeler => {
+  /* Register basic node types */
+  for (const node of nodeTypes) {
+    modeler.registerNode(node);
+  }
+
+  /* Add a BPMN extension */
+  modeler.registerBpmnExtension('pm', bpmnExtension);
+});
+
+window.ProcessMaker.EventBus.$on('modeler-start', modeler => {
+  modeler.loadXML(blank);
+});
+
